Add initial support for `STAGGER` pattern
It is not very usable yet because without defining `extra_space` (which
is not supported with plugin GUI) the resulting pattern can't be really
routed. There has to be an extra space in order to fit tracks in between
vias. Support for `extra_space` calculation based on track will be added
later.